backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pydantic
pydantic-settings
fastapi-cors
